+++ src/main/java/open/commons/spring/web/mvc/service/AbstractGenericService.java
@@ -26,8 +26,14 @@
 
 package open.commons.spring.web.mvc.service;
 
+import java.util.List;
+import java.util.function.Function;
+
 import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.InitializingBean;
+
+import open.commons.Result;
+import open.commons.function.TripleFunction;
 
 /**
  * 
@@ -35,7 +41,7 @@ import org.springframework.beans.factory.InitializingBean;
  * @version
  * @author Park_Jun_Hong_(parkjunhong77@gmail.com)
  */
-public abstract class AbstractGenericService extends AbstractComponent implements InitializingBean, DisposableBean {
+public abstract class AbstractGenericService extends AbstractComponent implements InitializingBean, DisposableBean, IConvertingService {
 
     /**
      * <br>
@@ -71,5 +77,59 @@ public abstract class AbstractGenericService extends AbstractComponent implement
      */
     @Override
     public void destroy() throws Exception {
+    }
+
+    /**
+     * 
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜      | 작성자   |   내용
+     * ------------------------------------------
+     * 2021. 12. 6.     박준홍         최초 작성
+     * </pre>
+     *
+     * @param <E>
+     *            Table Entity 타입
+     * @param <D>
+     *            DTO 타입
+     * @param <P>
+     *            파라미터 타입
+     * @param type
+     *            검색 유형.
+     * @param funcAll
+     *            전체 검색 함수
+     * @param param
+     *            검색 파라미터
+     * @param funcPagination
+     *            Pagination 검색 함수
+     * @param page
+     *            볼 페이지 번호 (1부터 시작)
+     * @param pageSize
+     *            1 페이지당 데이터 개수
+     * @param dtoType
+     *            DTO class
+     * @param converter
+     *            Entity -> DTO 변환 함수
+     * @return
+     *
+     * @since 2021. 12. 6.
+     * @version 0.4.0
+     * @author Park_Jun_Hong_(fafanmama_at_naver_com)
+     */
+    protected <E, D, P> Result<List<D>> selectMulti(SearchResultType type //
+            , Function<P, Result<List<E>>> funcAll //
+            , P param //
+            , TripleFunction<P, Integer, Integer, Result<List<E>>> funcPagination, int page, int pageSize //
+            , Class<D> dtoType, Function<E, D> converter) {
+        switch (type) {
+            case ALL:
+                return convertMultiResult(funcAll.apply(param), dtoType, converter);
+            case PAGINATION:
+                return convertMultiResult(funcPagination.apply(param, (page - 1) * pageSize, pageSize), dtoType, converter);
+            default:
+                throw new UnsupportedOperationException(String.format("지원하지 않음. type=%s", type.get()));
+        }
     }
 }